+++ src/components/layout/Breadcrumbs.tsx
@@ -20,6 +20,9 @@ const defaultLabels: RouteLabels = {
   "/yllapito": { label: "Etusivu", hideWhenNotCurrentRoute: true },
   "/yllapito/perusta": { label: "Projektin perustaminen" },
   "/yllapito/projekti": { label: "Projektit" },
+  "/yllapito/projekti/[oid]": { label: "Projekti..." },
+  "/yllapito/projekti/[oid]/aloituskuulutus": { label: "Aloituskuulutus" },
+  "/yllapito/projekti/[oid]/henkilot": { label: "Henkilöt ja käyttöoikeushallinta" },
   "/yllapito/ohjeet": { label: "Ohjeet" },
   "/suunnitelma": { label: "Suunnitelmat" },
   "/suunnitelma/[...all]": { label: "Tutki suunnitelmaa" },
