+++ src/pages/yllapito/index.tsx
@@ -1,12 +1,12 @@
 import { ProjektiListaus } from "@components/projektiListaus";
-import Button from "@components/button/Button";
+import ButtonLink from "@components/button/ButtonLink";
 
 function listingPage() {
   return (
     <section>
-      <Button primary link={{ href: "/yllapito/perusta" }}>
+      <ButtonLink primary href="/yllapito/perusta">
         Perusta uusi projekti
-      </Button>
+      </ButtonLink>
       <h2 className="mt-4">Projektit</h2>
       <ProjektiListaus admin />
     </section>
