+++ src/pages/yllapito/projekti/[oid]/nahtavillaolo.tsx
@@ -1,0 +1,31 @@
+import React, { ReactElement, useEffect } from "react";
+import ProjektiPageLayout from "@components/projekti/ProjektiPageLayout";
+import { useRouter } from "next/router";
+import useProjekti from "src/hooks/useProjekti";
+import { PageProps } from "@pages/_app";
+
+export default function Nahtavillaolo({ setRouteLabels }: PageProps): ReactElement {
+  const router = useRouter();
+  const oid = typeof router.query.oid === "string" ? router.query.oid : undefined;
+  const { data: projekti } = useProjekti(oid);
+
+  useEffect(() => {
+    if (router.isReady) {
+      let routeLabel = "";
+      if (projekti?.velho?.nimi) {
+        routeLabel = projekti.velho?.nimi;
+      } else if (typeof oid === "string") {
+        routeLabel = oid;
+      }
+      if (routeLabel) {
+        setRouteLabels({ "/yllapito/projekti/[oid]": { label: routeLabel } });
+      }
+    }
+  }, [router.isReady, oid, projekti, setRouteLabels]);
+
+  return (
+    <ProjektiPageLayout title="Nähtävilläolovaihe">
+      <div></div>
+    </ProjektiPageLayout>
+  );
+}
