+++ src/styles/button/ButtonWrapper.module.css
@@ -1,0 +1,3 @@
+.wrapper {
+  @apply relative inline-block;
+}
